pyproject.toml
[tool.poetry]
name = "discord-playlist-bot"
version = "0.1.0"
description = ""
authors = ["sarvagya14503 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.2"
flask = "1.1.4"
ytmusicapi = "^0.17.3"
youtube-dl = "^2021.5.16"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
